+++ src/app/app.component.ts
@@ -1,11 +1,9 @@
-import { Component, ViewChild, OnInit } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 import { Title, Meta } from '@angular/platform-browser';
 import { Router, ActivatedRoute, NavigationEnd } from '@angular/router';
 
 import { Subscription } from 'rxjs/Subscription';
 import { filter, map, switchMap } from 'rxjs/operators';
-
-import { CartListComponent } from './cart/components';
 
 @Component({
   selector: 'app-root',
@@ -13,10 +11,7 @@ import { CartListComponent } from './cart/components';
   styleUrls: ['./app.component.scss']
 })
 export class AppComponent implements OnInit {
-  @ViewChild('cart') cartListComponent: CartListComponent;
-
   private sub: Subscription;
-  cartFlag: false;
 
   constructor(
     private route: ActivatedRoute,
@@ -27,13 +22,6 @@ export class AppComponent implements OnInit {
 
   ngOnInit() {
     this.setPageTitlesAndMeta();
-  }
-
-  onAdded(flag) {
-    if (this.cartFlag) {
-      this.cartListComponent.onCartUpdate();
-    }
-    this.cartFlag = flag;
   }
 
   private setPageTitlesAndMeta() {