+++ src/app/core/+store/cart/index.ts
@@ -1,4 +1,5 @@
 export * from './cart.actions';
 export * from './cart.effects';
 export * from './cart.reducer';
+export * from './cart.selectors';
 export * from './cart.state';