+++ src/app/core/+store/products/products.reducer.ts
@@ -1,45 +1,31 @@
 import { productAdapter, ProductsState, initialProductsState } from './products.state';
-import { ProductsActionTypes } from './products.actions';
+import { ProductsActionTypes, ProductsActions } from './products.actions';
 
 import { IProduct } from '../../../shared/interfaces';
 
 export function productsReducer(
   state = initialProductsState,
-  action: ProductsState
+  action: ProductsActions
 ): ProductsState {
 
   switch (action.type) {
 
     case ProductsActionTypes.GET_PRODUCTS: {
-      return {
-        ...state,
-      };
+      return { ...state };
     }
 
     case ProductsActionTypes.GET_PRODUCTS_SUCCESS: {
-      const products = [...<Array<IProduct>>action.payload];
-
+      const products = [...<IProduct[]>action.payload];
       return productAdapter.addAll(products, {...state});
-    }
-
-
-    case ProductsActionTypes.GET_PRODUCTS_ERROR: {
-      const error = action.payload;
-      return {
-        ...state,
-        error
-      };
     }
 
     case ProductsActionTypes.CREATE_PRODUCT_SUCCESS: {
       const product = { ...<IProduct>action.payload };
-
       return productAdapter.addOne(product, state);
     }
 
     case ProductsActionTypes.UPDATE_PRODUCT_SUCCESS: {
       const product = { ...<IProduct>action.payload };
-
       return productAdapter.updateOne({
           id: product.id,
           changes: product
@@ -48,17 +34,17 @@ export function productsReducer(
 
     case ProductsActionTypes.DELETE_PRODUCT_SUCCESS: {
       const id = action.payload;
-
       return productAdapter.removeOne(id, state);
     }
 
+    case ProductsActionTypes.GET_PRODUCTS_ERROR:
     case ProductsActionTypes.CREATE_PRODUCT_ERROR:
     case ProductsActionTypes.UPDATE_PRODUCT_ERROR:
     case ProductsActionTypes.DELETE_PRODUCT_ERROR: {
       const error = action.payload;
+      console.log(`${action.type} error: `, error);
       return {
         ...state,
-        error
       };
     }