+++ src/app/core/+store/products/products.state.ts
@@ -3,11 +3,8 @@ import { createEntityAdapter, EntityState, EntityAdapter } from '@ngrx/entity';
 import { IProduct } from '../../../shared/interfaces';
 
 export interface ProductsState extends EntityState<IProduct> {
-  error: Error | string;
 }
 
 export const productAdapter: EntityAdapter<IProduct> = createEntityAdapter<IProduct>();
 
-export const initialProductsState: ProductsState = productAdapter.getInitialState({
-  error: null
-});
+export const initialProductsState: ProductsState = productAdapter.getInitialState();