+++ src/app/core/+store/products/produts.selectors.ts
@@ -1,5 +1,5 @@
 import { createFeatureSelector, createSelector } from '@ngrx/store';
-import { getRouterState } from './../router';
+import { getRouterState } from '../router';
 import { productAdapter, ProductsState } from './products.state';
 
 import { IProduct, Product } from '../../../shared/interfaces';
@@ -8,10 +8,8 @@ export const getProductsState = createFeatureSelector<ProductsState>('products')
 
 export const {
   selectEntities: getProductsEntities,
-  selectAll: getTasksData
+  selectAll: getProductsData
 } = productAdapter.getSelectors(getProductsState);
-
-export const getProductsError = createSelector(getProductsState, (state: ProductsState) => state.error);
 
 export const getSelectedProductByUrl = createSelector(
   getProductsEntities,